UI Element Under Mouse Picker
Editor extension that will create a new window and allow you to pick the UI element that is directly under the mouse cursor in the Scene window. Do this using the 1 key on your keyboard. 
The selection process will prioritise Z/distance order for picking the element over any other factor. It's only when the Z/distance is the same is when we have to try and figure out which item to select based on hierarchy position and sibling order - like the renderer does to choose which UI element to draw in front of other UI elements.